requirements.txt
# Web 框架
flask>=3.0.0
requests>=2.31.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0

# 数据分析和可视化（用于地址聚类分析）
networkx>=3.0.0
matplotlib>=3.7.0
numpy>=1.24.0

# 图像处理（用于聚类图生成）
pillow>=10.0.0

# 时间处理
python-dateutil>=2.8.0

# 配置管理
python-dotenv>=1.0.0

# 数据库
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0